tested (ShortestJobFirst)

+++ src/SchedulingAlgorithm.java
@@ -2,6 +2,6 @@ public interface SchedulingAlgorithm {
     public void getExecutionOrder();
     public void getWaitingTime();
     public void getTurnAroundTime();
-    public int getAverageWaitingTime();
-    public int getAverageTurnAroundTime();
+    public double getAverageWaitingTime();
+    public double getAverageTurnAroundTime();
 }
